Ten years of Rust! Happy birthday! 🎉
This is the first Rust program I ever wrote, convinced that the best way
to learn this hot new language of the future is by going through the
entire Project Euler problem archive.

Then I got stuck on trying to find prime numbers.

Can't remember if the language was in alpha or pre-alpha. I only
remember that it was winter, and that I had just discovered the
Solarized color scheme. That's well over a decade of not knowing how to
write software!